feat(FinishedJobs): add feature flag; fix circular deps
