+++ packages/javascript/bh-shared-ui/src/hooks/index.ts
@@ -14,15 +14,15 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-export { default as useOnClickOutside } from './useOnClickOutside';
-
-export { default as useDebouncedValue } from './useDebouncedValue';
-
-export { default as useToggle } from './useToggle';
-
 export { default as useApiVersion } from './useApiVersion';
 
 export { default as useCreateDisableZoomRef } from './useCreateDisableZoomRef';
+
+export { default as useDebouncedValue } from './useDebouncedValue';
+
+export { default as useOnClickOutside } from './useOnClickOutside';
+
+export { default as useToggle } from './useToggle';
 
 export * from './useAssetGroupTags';
 
@@ -30,7 +30,19 @@ export * from './useAvailableEnvironments';
 
 export * from './useConfiguration';
 
+export * from './useCustomNodeKinds';
+
 export * from './useDataQualityStats';
+
+export * from './useEnvironmentParams';
+
+export * from './useExploreGraph';
+
+export * from './useExploreParams';
+
+export * from './useExploreSelectedItem';
+
+export * from './useExploreTableAutoDisplay';
 
 export * from './useFeatureFlags';
 
@@ -38,40 +50,30 @@ export * from './useFetchEntityProperties';
 
 export * from './useFileIngest';
 
+export * from './useFinishedJobs';
+
+export * from './useGraphHasData';
+
 export * from './useGraphItem';
 
-export * from './useMountEffect';
-
-export * from './usePermissions';
-
-export * from './useSavedQueries';
-
-export * from './useSearch';
-
-export * from './useExploreParams';
-
-export * from './useExploreGraph';
-
-export * from './useExploreSelectedItem';
-
-export * from './useEnvironmentParams';
-
-export * from './useShowNavBar';
+export * from './useInitialEnvironment';
 
 export * from './useIsMouseDragging';
 
 export * from './useMatchingPaths';
 
-export * from './useInitialEnvironment';
+export * from './useMountEffect';
+
+export * from './usePermissions';
 
 export * from './usePreviousValue';
 
-export * from './useCustomNodeKinds';
+export * from './useSavedQueries';
 
-export * from './useExploreTableAutoDisplay';
-
-export * from './useGraphHasData';
+export * from './useSearch';
 
 export * from './useSelectedTagName';
+
+export * from './useShowNavBar';
 
 export * from './useZoneParams';
